requirements.txt
streamlit
google-generativeai
speechrecognition
sounddevice
numpy
gtts
scipy
